# product-requester-api/src/main/java/com/kh0ma/product_requester_api/dto/ProductCategoryDto.java
package com.kh0ma.product_requester_api.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import com.google.common.base.MoreObjects;

public class ProductCategoryDto {
    @JsonProperty("name")
    private String name;
    @JsonProperty("description")
    private String description;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    @Override
    public String toString() {
        return MoreObjects.toStringHelper(this)
                .add("name", name)
                .add("description", description)
                .toString();
    }
}